feat: contract, account, blockchain -> core as dep
