+++ packages/zilliqa-js-blockchain/tsconfig.json
@@ -7,7 +7,6 @@
   "include": ["src", "../../typings/**/*.d.ts"],
   "references": [
     { "path": "../zilliqa-js-account" },
-    { "path": "../zilliqa-js-contract" },
     { "path": "../zilliqa-js-core" }
   ]
 }
